backend/requirements.txt
fastapi
uvicorn
python-dotenv
os
tempfile
traceback
pydantic
